Abstracted out websocket interactions into an abstract base class that OrderBook now inherits from. This allows easy creation of new types of listeners for the GDAX websocket API.
*****************************
*** WebSocketFeedListener ***
*****************************
Factored the web socket code away from the actual orderbook
implementation. The new WebSocketFeedListener is an abstract base class
that handles the websocket itself including:

- Connect
- Disconnect
- Send / Receive messages
- Subscribing to channels
- Leveraging the trade log file
- Async Context Manager calls

WebSocketFeedListener inherits from ABC and has a single @abstractmethod
handle_message(self), which must be implemented by subclasses.

******************************
********* OrderBook **********
******************************

Removed all WebSocket manipulation from OrderBook and left just the
OrderBook implementation. OrderBook now inherits from
WebSocketFeedListener and implements handle_message.
