+++ gdax/__init__.py
@@ -1,3 +1,4 @@
 import gdax.orderbook
 import gdax.trader
 import gdax.utils
+import gdax.websocket_feed_listener
